Better encapsulation
No more leaking ofMove PatchedFiles in a settings option. Move all
patch classes into a `Rewrites` object.
